Updated code structure.

+++ DiophAppr.cpp
@@ -2,7 +2,7 @@
 
 namespace Elementary
 {
-	std::tuple<int64_t, int64_t, int64_t, int> DiophAppr::ApprByRedFrac(Fraction alpha, int64_t Q)
+	std::tuple<int128_t, int128_t, int128_t, int> DiophAppr::ApprByRedFrac(Fraction alpha, int64_t Q)
 	{
 		int sgn_alpha = 1;
 		if (alpha.IsNegative())
@@ -11,15 +11,15 @@ namespace Elementary
 			alpha.Negate();
 		}
 
-		int64_t p[2]{ 0, 1 };
-		int64_t q[2]{ 1, 0 };
+		int128_t p[2]{ 0, 1 };
+		int128_t q[2]{ 1, 0 };
 		int s = 1;
 
 		while (q[1] <= Q)
 		{
-			int64_t a = alpha.Floor();
-			int64_t new_p = a * p[1] + p[0];
-			int64_t new_q = a * q[1] + q[0];
+			int128_t a = alpha.Floor();
+			int128_t new_p = a * p[1] + p[0];
+			int128_t new_q = a * q[1] + q[0];
 			p[0] = p[1]; p[1] = new_p;
 			q[0] = q[1]; q[1] = new_q;
 
@@ -27,7 +27,9 @@ namespace Elementary
 			{
 				return std::make_tuple(sgn_alpha * p[1], (-sgn_alpha * s * q[0]) % q[1], q[1], 0);
 			}
-			alpha = alpha.FractionalPart(); alpha.Invert(); // alpha <- 1/(alpha - a).
+
+			alpha = alpha.FractionalPart();
+			alpha.Invert();
 			s *= -1;
 		}
 		// As q[1] > Q, we resort to the previous approximation, i.e. pm / qm.
